inference-service/requirements.txt
# ML Framework & Model
torch>=1.12     # PyTorch version with MPS support
torchvision
torchaudio
ultralytics     # For YOLOv8 models and utilities

# Image Processing
opencv-python-headless>=4.5  # Use headless if no GUI needed in container
numpy>=1.23

# Message Queue
pika>=1.3

# Utilities & Config (Optional but good practice)
python-dotenv>=1.0.0 # For loading environment variables from .env file
pyyaml>=6.0          # If using YAML config files

# Optional: For health checks/metrics API
# fastapi>=0.95
# uvicorn>=0.23
# prometheus-client>=0.15.0